Add unit tests for MultiTE and T2Scalar ASL mapping functionalities
- Created `test_multi_te_mapping.py` to test MultiTE_ASLMapping class, including methods for setting brain masks, CBF and ATT maps, and creating maps.
- Implemented tests for error handling when ASLData is incomplete or when invalid parameters are provided.
- Added `test_te_mapping.py` for T2Scalar_ASLMapping class, verifying initialization, error handling for missing TE and PLD values, and successful T2 map creation.
- Removed the outdated `test_reconstruction.py` file to streamline test organization and improve maintainability.
