+++ asltk/reconstruction/__init__.py
@@ -1,6 +1,11 @@
-from .t2_mapping import T2Scalar_ASLMapping
 from .cbf_mapping import CBFMapping
 from .multi_dw_mapping import MultiDW_ASLMapping
 from .multi_te_mapping import MultiTE_ASLMapping
+from .t2_mapping import T2Scalar_ASLMapping
 
-__all__ = ['CBFMapping', 'MultiTE_ASLMapping', 'MultiDW_ASLMapping', 'T2Scalar_ASLMapping']
+__all__ = [
+    'CBFMapping',
+    'MultiTE_ASLMapping',
+    'MultiDW_ASLMapping',
+    'T2Scalar_ASLMapping',
+]
